Add taks status

+++ src/lib/formatters.ts
@@ -12,7 +12,7 @@ export function formatDate(isoDateStr: string) {
   const input = new Date(year, month - 1, day);
 
   const dayStr = input.getDate();
-  let monthStr = input.toLocaleString('es-NI', { month: 'long' });
+  let monthStr = input.toLocaleString('es-NI', { month: 'short' });
   const yearStr = input.getFullYear();
 
   const currentYear = new Date().getFullYear();
